gate pass done

+++ src/main/java/lk/zerocode/api/controller/response/OtherLeavesResponse.java
@@ -43,5 +43,5 @@ public class OtherLeavesResponse {
 
     private Status status;
 
-    private Integer hours;
+    private Float hours;
 }
